Finished with project
this is my last commit,

The overall project went well, I overscoped a bit because I wanted to challenge myself to learn more.

Scriptable objects was initially a problem with the new UI builder but its good that changed in UIbuilder (editorwindow) counts as editing serialized fields which means scriptable objects are saved.

The code could be a lot cleaner, and if I had a extra week it could be improved upon, both from a readability point of view, but also add more to the tool and get a deeper understanding of the new UI builder.